Add option to embed PDF in e-Invoice and improve ANAF e-Invoice import
This update introduces a new company setting that allows the option to embed the PDF invoice in the Romanian e-Invoice (CIUS-RO). It also refactors the ANAF e-Invoicing processing mechanisms in the company settings to improve the import procedure from ANAF. Furthermore, necessary adjustments are made related to handling of invoice downloads and transactions.